Разделите код, уже написанный в main.js, на отдельные ES2015-модули. main.js станет главным модулем, точкой входа. В этом файле должны остаться только импорты других модулей, а также код, который необходим для их работы.
Пример разбиения на модули
Импортируйте нужные модули в точку входа, а точку входа подключите в index.html как скрипт-модуль.

По желанию, вы можете перечитать техническое задание и подумать, какие ещё модули потребуются в приложении, а после завести под них пустые файлы, чтобы сразу обозначить файловую структуру проекта.
