+++ js/Mock/InitialData.js
@@ -1,0 +1,44 @@
+export const initialData = {
+  countPreview: {
+    start: 1,
+    end: 25,
+  },
+  names: [
+    `Анна`,
+    `Мария`,
+    `Александр`,
+    `Влад`,
+    `Константин`,
+    `Егор`,
+    `Милена`,
+    `Эсмеральда`,
+  ],
+  comments: {
+    count: {
+      min: 0,
+      max: 135,
+    },
+    avatar: {
+      avatarMin: 1,
+      avatarMax: 6,
+    },
+    message: {
+      length: {
+        min: 1,
+        max: 2,
+      },
+      message: [
+        `Всё отлично!`,
+        `В целом всё неплохо. Но не всё.`,
+        `Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.`,
+        `Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.`,
+        `Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.`,
+        `Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!`,
+      ],
+    },
+  },
+  likes: {
+    min: 15,
+    max: 200,
+  },
+};
